api_rest/pyproject.toml
[build-system]
requires = ["setuptools>=67.0", "setuptools_scm[toml]>=7.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
readme = { file = ["README.md"] }

[tool.setuptools.package-data]
api_rest = ["py.typed"]

[tool.setuptools_scm]
fallback_version = "0.0.0-dev"
write_to = "api_rest/_version.py"

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*"]

[project]
authors = [{ "name" = "Andres" }]
description = "Ejercicio"
dynamic = ["version"]
license = { "file" = "LICENSE" }
name = "api_rest"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"

dependencies = [
  "SQLAlchemy",
  "aiosqlite",
  "alembic",
  "asyncpg",
  "fastapi",
  "psycopg2-binary",
  "pydantic[email]~=2.0",
  "pydantic-settings",
  "uvicorn[standard]",
  "python-jose[cryptography]>=3.5.0",
  "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
  "build",
  "dapperdata",
  "glom",
  "greenlet",
  "httpx",
  "mypy",
  "paracelsus",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-pretty",
  "ruamel.yaml",
  "ruff",
  "tombi",
  "uv"
]

[tool.ruff]
exclude = [".venv", "./api_rest/_version.py"]
line-length = 120

[tool.coverage.run]
concurrency = ["thread", "greenlet"]
omit = [
  "./api_rest/_version.py",
  "./api_rest/__init__.py",
  "./tests/*",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
strict_optional = true
warn_return_any = true
warn_unused_ignores = true

[tool.uv]
exclude-newer = "1 week"
